requirements.txt
click==8.1.2
Flask==1.1.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==1.1.0
lyricsgenius==3.0.1
MarkupSafe==1.1.1
nltk==3.7
numpy==1.21.5
pandas==0.24.2
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.1.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
spotipy==2.19.0
SQLAlchemy==1.4.35
textblob==0.17.1
threadpoolctl==3.1.0
typing-extensions==4.1.1
Werkzeug==0.15.4
zipp==3.8.0
